+++ test/unit/Base.php
@@ -6,7 +6,7 @@
  * @author    Coby Tamayo <ctamayo@sitecrafting.com>
  */
 
-namespace ConiferTest;
+namespace Conifer\Unit;
 
 use PHPUnit\Framework\TestCase;
 use Timber\User;
