+++ test/unit/TestHelperTest.php
@@ -6,7 +6,7 @@
  * @author    Coby Tamayo <ctamayo@sitecrafting.com>
  */
 
-namespace ConiferTest;
+namespace Conifer\Unit;
 
 use Conifer\Twig\TextHelper;
