+++ test/unit/Twig/Filters/FormHelperTest.php
@@ -6,7 +6,7 @@
  * @author Coby Tamayo
  */
 
-namespace ConiferTest;
+namespace Conifer\Unit;
 
 use Conifer\Form\AbstractBase as Form;
 use Conifer\Twig\FormHelper;
